Source/backend/requirements.txt
# Workspace version: 1.44.15
databutton==0.38.34
ipykernel==6.29.5
fastapi==0.111.0
python-multipart==0.0.9
uvicorn[standard]==0.29.0

openai
beautifulsoup4
requests
line-bot-sdk
anthropic